cloud-function/requirements.txt
google-cloud-storage
pytest    
ipython         
